Commit/Sync improvements
