+++ FitpaySDK/PaymentDevice/Operations/SyncOperationError.swift
@@ -1,0 +1,10 @@
+import Foundation
+
+enum SyncOperationError: Error {
+    case deviceIdentifierIsNil
+    case couldNotConnectToDevice
+    case paymentDeviceDisconnected
+    case cantFetchCommits
+    case alreadySyncing
+    case unk
+}
